feat(ai): add AI configuration persistence and MCP commands
- Add ai_config table (migration 000014)
- Implement gorev_ai_config MCP tool (get/set/test actions)
- Add AI config CRUD operations to veri_yonetici
- Update DecomposeTask, AnalyzeProject, EstimateTime to use extractJSON()
- Enhance HTTP client with retry logic and timeout handling
- Add LoadConfigFromDB helper for reloading config from database

Database changes:
- ai_config table with provider, api_key, model, temperature, max_tokens
- Supports per-project AI configuration
- Automatic provider registration on config load

MCP tools:
- gorev_ai_config action=get: Retrieve AI configuration
- gorev_ai_config action=set: Save AI configuration (provider, api_key, model)
- gorev_ai_config action=test: Test AI provider connection

+++ gorev-mcpserver/internal/ai/providers/provider.go
@@ -167,8 +167,8 @@ func (p *BaseProvider) ModelCost(modelID string, usage ChatUsage) (float64, erro
 	}
 
 	// Calculate cost
-	inputCost := float64(usage.PromptTokens) / 1000000 * pricing.Input
-	outputCost := float64(usage.CompletionTokens) / 1000000 * pricing.Output
+	inputCost := float64(usage.PromptTokens) / 1000000 * float64(pricing.Input)
+	outputCost := float64(usage.CompletionTokens) / 1000000 * float64(pricing.Output)
 	totalCost := inputCost + outputCost
 
 	return totalCost, nil
